refactor: move entities to infrastructure-module to seperate it from domain-module
