+++ domain/src/main/java/hanghae/domain/types/showtime/Schedule.java
