+++ domain/src/main/java/hanghae/domain/types/theater/ShowtimeCollection.java
