+++ infrastructure/src/main/java/hanghae/infrastructure/domain/entity/TheaterShowtimeEntity.java
@@ -1,22 +1,22 @@
-package hanghae.domain.entity;
+package hanghae.infrastructure.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "movie_theater")
+@Table(name = "theater_showtime")
 @NoArgsConstructor
-public class MovieTheaterEntity {
+public class TheaterShowtimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long movieTheaterId;
-
-    @ManyToOne
-    @JoinColumn(name = "movie_id", nullable = false)
-    private MovieEntity movie;
+    private Long theaterShowtimeId;
 
     @ManyToOne
     @JoinColumn(name = "theater_id", nullable = false)
     private TheaterEntity theater;
+
+    @ManyToOne
+    @JoinColumn(name = "showtime_id", nullable = false)
+    private ShowtimeEntity showtime;
 }
