+++ infrastructure/src/main/java/hanghae/infrastructure/domain/mixin/ScheduleMixin.java
@@ -1,0 +1,2 @@
+package hanghae.infrastructure.domain.mixin;public class ScheduleMixin {
+}
