+++ infrastructure/src/main/java/hanghae/infrastructure/repository/JpaMovieTheaterRepository.java
@@ -1,0 +1,4 @@
+package hanghae.infrastructure.repository;
+
+public interface MovieTheaterRepository {
+}
